n to n relationship CategoryGame

+++ Controllers/CategoryController.cs
@@ -9,9 +9,13 @@ namespace GameLibrary.Controllers
     public class CategoryController : Controller
     {
         private readonly ICategoryServices _catservices;
-        public CategoryController(ICategoryServices catservices)
+        private readonly IGameServices _gameServices;
+        private readonly ICategoryGameServices _categoryGameServices;
+        public CategoryController(ICategoryServices catservices, ICategoryGameServices categoryGameServices, IGameServices gameServices)
         {
             _catservices = catservices;
+            _categoryGameServices = categoryGameServices;
+            _gameServices = gameServices; 
         }
 
         public async Task<IActionResult> Index()
@@ -24,7 +28,7 @@ namespace GameLibrary.Controllers
         public async Task<IActionResult> Create()
         {
             CategoryForCreationDto categoryForCreationDto = new CategoryForCreationDto();
-
+            categoryForCreationDto.Games = await _gameServices.GetAllGames();
             return View(categoryForCreationDto);
         }
         
@@ -43,14 +47,33 @@ namespace GameLibrary.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
-            var itemToDelete = await _catservices.GetCategoryByIdAsync(id);
-            await _catservices.DeleteCategoriesAsync(itemToDelete);
-            return RedirectToAction("Index");
+            var itemToDelete = await _catservices.DetailsCategoryByid(id);
+
+            var allRelationships = await _categoryGameServices.GetAllCategoryGamesAsync();
+
+
+
+            if (itemToDelete.CategoryGames != null && itemToDelete.CategoryGames.Any())
+            {
+                return RedirectToAction("Index");
+            }
+            else if (itemToDelete.CategoryGames == null || itemToDelete.CategoryGames.Count == 0 || itemToDelete.CategoryGames.All(x => x == null))
+            {
+                await _catservices.DeleteCategoriesAsync(itemToDelete);
+
+
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index");
+            }
+
         }
 
         public async Task<IActionResult> Details(int id)
         {
-            var category = await _catservices.GetCategoryByIdAsync(id);
+            var category = await _catservices.DetailsCategoryByid(id);
             return View(category);
         }
